week1/requirements.txt
nicegui
google-cloud-aiplatform
langchain>=0.0.142
langchain_openai
